f2c_create_db/parser/ErrorListener: token types are signed
Esp. EOF is -1. So treat them as ssize_t.

+++ f2c_create_db/parser/ErrorListener.cpp
@@ -83,7 +83,7 @@ void ErrorListener::parserError(antlr4::Parser &parser, antlr4::Token &offending
 	auto tokEnd = std::min(tokStart + surround * 2, tokens->size() - 1);
 	SlHelpers::String::join(std::cerr, tokens->get(tokStart, tokEnd),
 				[](auto &out, const auto &tok) {
-		out << '[' << tok->getType() << ",'";
+		out << '[' << static_cast<ssize_t>(tok->getType()) << ",'";
 		for (char c: tok->getText())
 			switch (c) {
 			case '\t':
